Deletenado uma tarefa, realizando autenticação com sanctum, protegendo rotas com sanctum

+++ routes/api.php
@@ -3,6 +3,8 @@
 use Illuminate\Http\Request;
 use Illuminate\Support\Facades\Route;
 use App\Http\Controllers\ApiController;
+use App\Http\Controllers\AuthController;
+use Illuminate\Support\Facades\Auth;
 
 /*
 |--------------------------------------------------------------------------
@@ -21,9 +23,18 @@ Route::get('/ping', function(){
     ];
 });
 
-Route::post('/todo', [ApiController::class, 'createtodo']);
+Route::get('/unauthenticated', function(){
+    return ['error' => 'usuário não logado'];
+})->name('login');
+
+Route::post('/user', [AuthController::class, 'create']);
+Route::middleware('auth:sanctum')->get('/auth/logout', [AuthController::class, 'logout']);
+Route::post('/auth', [AuthController::class, 'login']);
+
+
+Route::middleware('auth:sanctum')->post('/todo', [ApiController::class, 'createtodo']);
 Route::get('/todos', [ApiController::class, 'readalltodos']);
 Route::get('todo/{id}', [ApiController::class, 'readtodo']);
-Route::put('todo/{id}', [ApiController::class, 'updatetodo']);
-Route::delete('/todo/{ìd}', [ApiController::class, 'deletetodo']);
+Route::middleware('auth:sanctum')->put('todo/{id}', [ApiController::class, 'updatetodo']);
+Route::middleware('auth:sanctum')->delete('/todo/{ìd}', [ApiController::class, 'deletetodo']);
